split imeplementation to allow exporting trait
