+++ Cargo.toml
@@ -13,13 +13,13 @@ repository = "https://github.com/lesurp/OptionalStruct"
 exclude = [".travis.yml"]
 edition = '2021'
 
+[workspace]
+members = ["internal", "macro_impl"]
+
 [badges]
 travis-ci = { repository = "lesurp/OptionalStruct", branch = "master" }
 
 [dependencies]
-syn = { version = "1.0.107", features = ["parsing", "proc-macro"] }
-quote = "1.0"
-proc-macro2 = "1.0.49"
+internal = { path = "internal" }
+macro_impl = { path = "macro_impl" }
 
-[lib]
-proc-macro = true
