+++ macro_impl/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "macro_impl"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+syn = { version = "1.0.107", features = ["parsing", "proc-macro"] }
+quote = "1.0"
+proc-macro2 = "1.0.49"
+internal = { path = "../internal" }
+
+[lib]
+proc-macro = true
